requirements.txt
# Core runtime
pyyaml>=6.0
python-dotenv>=1.0.0

# LLM APIs
openai>=1.0.0
google-generativeai>=0.3.0

# LangGraph
langgraph>=0.0.20
langchain>=0.1.0

# Retrieval
faiss-cpu>=1.7.4
rank-bm25>=0.2.2
numpy>=1.24.0
sentence-transformers>=2.2.0

# Medical entity extraction
# Note: medcat 2.x requires pydantic>=2.0 and spacy>=3.8
medcat>=2.0
spacy>=3.8
peft>=0.10.0

# Multilingual translation (Helsinki-NLP opus-mt)
transformers>=4.30.0
torch>=2.0.0
langdetect>=1.0.9
sentencepiece>=0.1.99

# Optional: Web UI
streamlit>=1.28.0

# Evaluation metrics
ragas>=0.1.0
datasets>=2.14.0

